[DRAFT] clean, add comments, Add origins to operations and replace Float by BigDecimal

+++ fr.rostren.tracker.model.edit/src-gen/fr/rostren/tracker/provider/DebitOperationItemProvider.java
@@ -2,6 +2,7 @@
  */
 package fr.rostren.tracker.provider;
 
+import java.math.BigDecimal;
 import java.util.Collection;
 import java.util.List;
 
@@ -17,9 +18,10 @@ import org.eclipse.emf.edit.provider.ITreeItemContentProvider;
 import fr.rostren.tracker.DebitOperation;
 
 /**
- * This is the item provider adapter for a {@link fr.rostren.tracker.DebitOperation} object.
- * <!-- begin-user-doc -->
+ * This is the item provider adapter for a
+ * {@link fr.rostren.tracker.DebitOperation} object. <!-- begin-user-doc -->
  * <!-- end-user-doc -->
+ * 
  * @generated
  */
 public class DebitOperationItemProvider extends OperationItemProvider implements
@@ -51,14 +53,15 @@ public class DebitOperationItemProvider extends OperationItemProvider implements
 	}
 
 	/**
-	 * This returns DebitOperation.gif.
-	 * <!-- begin-user-doc --> <!--
+	 * This returns DebitOperation.gif. <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
+	 * 
 	 * @generated
 	 */
 	@Override
 	public Object getImage(Object object) {
-		return overlayImage(object, getResourceLocator().getImage("full/obj16/DebitOperation"));
+		return overlayImage(object,
+				getResourceLocator().getImage("full/obj16/DebitOperation"));
 	}
 
 	/**
@@ -71,12 +74,14 @@ public class DebitOperationItemProvider extends OperationItemProvider implements
 	public String getText(Object object) {
 		String operationTitle = ((DebitOperation) object).getOperationTitle()
 				.getTitle();
-		String operationAmount = ((Float) ((DebitOperation) object)
+		String operationAmount = ((BigDecimal) ((DebitOperation) object)
 				.getTotalAmount()).toString();
 		if (operationTitle == null) {
 			return "New " + getString("_UI_DebitOperation_type");
 		} else if (operationTitle.length() == 0) {
-			return getString("_UI_DebitOperation_type") + " --> Undefined Operation Title = " + operationAmount + " euros";
+			return getString("_UI_DebitOperation_type")
+					+ " --> Undefined Operation Title = " + operationAmount
+					+ " euros";
 		} else {
 			return getString("_UI_DebitOperation_type") + " --> "
 					+ operationTitle + " = " + operationAmount + " euros";
@@ -84,10 +89,11 @@ public class DebitOperationItemProvider extends OperationItemProvider implements
 	}
 
 	/**
-	 * This handles model notifications by calling {@link #updateChildren} to update any cached
-	 * children and by creating a viewer notification, which it passes to {@link #fireNotifyChanged}.
-	 * <!-- begin-user-doc --> <!--
+	 * This handles model notifications by calling {@link #updateChildren} to
+	 * update any cached children and by creating a viewer notification, which
+	 * it passes to {@link #fireNotifyChanged}. <!-- begin-user-doc --> <!--
 	 * end-user-doc -->
+	 * 
 	 * @generated
 	 */
 	@Override
