+++ fr.rostren.tracker.model.edit/src/fr/rostren/tracker/provider/dev/IncomingTransferItemProviderDev.java
@@ -2,6 +2,7 @@
  */
 package fr.rostren.tracker.provider.dev;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
@@ -45,7 +46,7 @@ public class IncomingTransferItemProviderDev extends
 			return super.createSetCommand(domain, owner, feature, value, index);
 
 		final IncomingTransfer incomingTransfer = (IncomingTransfer) owner;
-		final float newTotalAmount = (Float) value;
+		final BigDecimal newTotalAmount = (BigDecimal) value;
 
 		if (incomingTransfer.getTotalAmount() == newTotalAmount)
 			return super.createSetCommand(domain, owner, feature, value, index);
